less warns

+++ core/patterndetection/chessBoardDetector.cpp
@@ -71,8 +71,8 @@ bool ChessboardDetector::detectPattern(DpImage &buffer)
 
     for (auto& b: boards)
     {
-        int bw = b[0].size();
-        int bh = b.size();
+        int bw = (int)b[0].size();
+        int bh = (int)b.size();
 
         bool fitw = (bw == horCrossesCount());
         bool fith = (bh == vertCrossesCount());
